# ArchitectureRefactoringHelper/Repository/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/approaches": {
      "get": {
        "tags": [
          "RefactoringApproach"
        ],
        "summary": "Receives a complete list of all RefactoringApproach items",
        "operationId": "ListRefactoringApproaches",
        "parameters": [
          {
            "name": "withDetails",
            "in": "query",
            "description": "Decides whether the approaches are returned with all details or not",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefactoringApproach"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "summary": "Creates a new RefactoringApproach.\r\nInputs, process attributes, outputs or usability attributes that don't exist will be created in the process.",
        "operationId": "AddRefactoringApproach",
        "requestBody": {
          "description": "The RefactoringApproach we want to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefactoringApproach"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefactoringApproach"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefactoringApproach"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created RefactoringApproach",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefactoringApproach"
                }
              }
            }
          },
          "400": {
            "description": "If the RefactoringApproach is null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "If a RefactoringApproach with the same title already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approaches/{id}": {
      "get": {
        "tags": [
          "RefactoringApproach"
        ],
        "summary": "Receives a RefactoringApproach",
        "operationId": "GetRefactoringApproach",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefactoringApproach"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "summary": "Deletes a RefactoringApproach",
        "operationId": "DeleteRefactoringApproach",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/recommendations": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "summary": "Returns a list of recommendation with their suitability evaluation based on either the request body or a selected\r\nrecommendation preset.\r\nIf a preset is given it will always be prioritised over the recommendation request body.",
        "operationId": "RecommendRefactoringApproaches",
        "parameters": [
          {
            "name": "preset",
            "in": "query",
            "description": "The recommendation preset the recommendation algorithm is running against",
            "schema": {
              "$ref": "#/components/schemas/RecommendationPreset"
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The number of recommendations returned. If the number is negative all elements are returned",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The recommendation information the recommendation algorithm is running against",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproachRecommendationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproachRecommendationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproachRecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApproachRecommendation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/domain-artifacts": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "AddDomainArtifactAsInput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainArtifactInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DomainArtifactInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DomainArtifactInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/domain-artifacts/{inputName}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveDomainArtifactFromInputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "inputName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/outputs/representations": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "AddRepresentationAsOutput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/outputs/representations/{outputName}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveRepresentationFromOutputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "outputName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/existing-cards": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveApproachExistingCards",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/runtime-artifacts": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "AddRuntimeArtifactAsInput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeArtifactInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeArtifactInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeArtifactInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/runtime-artifacts/{inputName}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveRuntimeArtifactFromInputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "inputName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/model-artifacts": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "AddModelArtifactAsInput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelArtifactInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelArtifactInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModelArtifactInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/model-artifacts/{inputName}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveModelArtifactFromInputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "inputName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/approaches/{id}/inputs/executables": {
      "post": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "AddExecutableAsInput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutableInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutableInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutableInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/approaches/{id}/processes/qualities/{qualityName}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveQualityFromProcess",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "qualityName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DomainArtifactInput": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RecommendationPreset": {
        "enum": [
          "NewApplication",
          "ReBuild",
          "ReFactor"
        ],
        "type": "string"
      },
...
      "Representation": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}